Apply fixes from StyleCI

+++ src/routes.php
@@ -1,6 +1,6 @@
 <?php
 
-if ((double) app()->version() >= 5.2) {
+if ((float) app()->version() >= 5.2) {
     Route::get('captcha', '\Hamog\Captcha\CaptchaController@getCaptcha')->middleware('web');
 } else {
     Route::get('captcha', '\Hamog\Captcha\CaptchaController@getCaptcha');
